port Kollect models to their own package and file
